Tests/LibWeb/Text/expected/HTML/hash-change-event.txt

type: hashchange
oldURL: https://example.com/old#🐞
newURL: https://example.com/new#🐞
